feat(back): in progress remove profile image

+++ back/src/photo-profile/resolvers/removeProfileImage.ts
@@ -1,20 +1,38 @@
 import { Resolver, Mutation, Args } from '@nestjs/graphql';
 import { PhotoProfileService } from '../photo-profile.service';
+import { CloudinaryService } from '../../utils/cloudinary.service';
+import { UseGuards, HttpException, HttpStatus } from '@nestjs/common';
+import { GqlAuthGuard, CurrentUser } from '../../auth/auth.guards';
+import { AuthPayload } from '../../auth/auth.model';
 
 @Resolver()
 export class RemoveProfileImageResolver {
-  constructor(private photoProfileService: PhotoProfileService) {}
+  constructor(
+    private photoProfileService: PhotoProfileService,
+    private cloudinaryService: CloudinaryService,
+  ) {}
 
   @Mutation(() => Boolean)
+  @UseGuards(GqlAuthGuard)
   async removeProfileImage(
-    @Args('idPhotoProfile') idPhotoProfile: number,
+    @Args('photoProfilePublicId') publicId: string,
+    @CurrentUser() authPayload: AuthPayload,
   ): Promise<boolean> {
-    const photoProfile = await this.photoProfileService.getPhotoProfileById(
-      idPhotoProfile,
+    const photoProfile = await this.photoProfileService.getPhotoProfileByPublicId(
+      publicId,
     );
-    const response = await this.photoProfileService.deletePhotoProfile(
+
+    if (photoProfile.idUser !== authPayload.payload.id)
+      throw new HttpException('CannotDeletePhoto', HttpStatus.UNAUTHORIZED);
+
+    const dbPhotoRemoved = await this.photoProfileService.deletePhotoProfile(
       photoProfile,
     );
-    return response && !response.id;
+    const cloudinaryResponse = await this.cloudinaryService.cloudinary.uploader.destroy(
+      publicId,
+    );
+    return (
+      dbPhotoRemoved && !dbPhotoRemoved.id && cloudinaryResponse.result === 'ok'
+    );
   }
 }
